led: [HAVE BUG] Optimize LED charging light state
By default, the LED deamon works just for a while, then exited, and soon
will be restarted. Now I make it as a real "deamon" --- use loop to
watch and change LED state.

But using while loop consumes too much CPU. So now I've to disable it.
The service should be rewritten.